+++ CHANGELOG.md
@@ -11,7 +11,10 @@
 - `PulseUpdate::getUpdates()` now correctly returns an array of PulseUpdate objects
 - `PulseUpdate::getUpdates()` now checks it's parameters and attempts to automatically convert them to a format DaPulse will accept
 - `PulseUpdate::likeUpdate()` and `PulseUpdate::unlikeUpdate()` now returns boolean values on whether it was successful or not; it also throws an InvalidArgumentException when the given parameter is not a valid user
-- `Pulse::editPulseName()` has been added in order to allow for one less API call than creating a Pulse object and then calling `editName()`
+- Added new functions to edit a Pulse without creating a Pulse object. If a Pulse object is not needed or is not already given, then these calls would be faster.
+    - `Pulse::editPulseNameByID()` = `Pulse::editName()`
+    - `Pulse::archivePulseByID()` = `Pulse::archivePulse()`
+    - `Pulse::deletePulseByID()` = `Pulse::deletePulse()`
 
 #### Deprecations